// front/src/providers/api-client.generated/model/jobOfferDto.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';
import { CustomerDto } from './customerDto';
import { CandidateApplicationDto } from './candidateApplicationDto';
import { AppValueDto } from './appValueDto';
import { CandidateApplicationJobsDto } from './candidateApplicationJobsDto';


export interface JobOfferDto { 
    id?: string;
    title: string;
    jobDescription: string;
    ref: string;
    creationDate?: Date;
    modifDate?: Date;
    city?: string;
    country?: string;
    countryCode?: string;
    consultantId?: string;
    consultant?: UserDto;
    salary?: string;
    publicLink?: string;
    jobId: string;
    job?: AppValueDto;
    taskResponsabilitiesDescription?: string;
    candidateProfileDescription?: string;
    conditionsDescription?: string;
    applyInCouple?: boolean;
    contractTypeId?: string;
    contractType?: AppValueDto;
    candidateApplicationJobs?: Array<CandidateApplicationJobsDto>;
    disabled?: boolean;
    customerId?: string;
    customer?: CustomerDto;
    stateId?: string;
    state?: AppValueDto;
    candidateApplications?: Array<CandidateApplicationDto>;
}